requirements.txt
openai
click
rich
python-dotenv
textual